chore(web): define shadcn preset with brand palette
